backend/requirements.txt
Django
djangorestframework
channels
daphne
python-chess
python-dotenv
django-cors-headers
